Move tests to test/jdk/tools/launcher/modules/patch/automatic.
