+++ test/jdk/tools/launcher/modules/patch/automatic/somelib/somelib/Dummy.java
